+++ BackEnd/controllers/users/index.js
@@ -1,0 +1,5 @@
+const { loginUser } = require('./loginController');
+const { idCheckUser } = require('./idCheckController');
+const { signupUser } = require('./signupController');
+
+module.exports = { loginUser, idCheckUser, signupUser };
