+++ BackEnd/routes/userRoutes.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const path = require('path');
 const userController = require(
-  path.join(__dirname, '../controllers/userController.js'),
+  path.join(__dirname, '../controllers/users/index.js'),
 );
 
 // user signup
